fix: include/linux/android_kabi.h: Fix ANDROID_KABI file
Ok so, reviewing and comparing the kabi files from the google's repo and
this one was wildly different.

Because there's a lot of duplicated lines in the source code. My
keyboard is to blame. I should have reviewed the copy before deployment
Which is also, i'm also at my fault

Signed-off-by: Mizumo-prjkt <bluepearl394@gmail.com>